client crud finished

+++ fishing-booker/src/main/java/com/isa/fishingbooker/controller/ClientController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -26,13 +26,11 @@ public class ClientController {
 	@Autowired 
 	private ClientService clientService;
 	
-	
 	@GetMapping("/clients")
 	public List<Client> getAllClients(){
 		return this.clientService.getAllClients();
 	}
 	
-
 	@GetMapping("/clients/{id}")
 	public ResponseEntity<Client> getClientById(@PathVariable(value= "id") int clientId)
 		throws ResourceNotFoundException{
@@ -44,19 +42,16 @@ public class ClientController {
 		return clientService.createClient(client);
 	}
 	
-	/*
 	@PutMapping("/clients/{id}")
-	public ResponseEntity<Client> updateClient(@PathVariable(value = "id") Long clientId,
+	public ResponseEntity<Client> updateClient(@PathVariable(value = "id") int clientId,
 			 @RequestBody Client clientDetails) throws ResourceNotFoundException {
 		return clientService.updateClient(clientId, clientDetails);
 	}
-	*/
+	
 	@DeleteMapping("/clients/{id}")
 	public Map<String, Boolean> deleteClient(@PathVariable(value = "id") int clientId)
 			throws ResourceNotFoundException {
 		return clientService.deleteClient(clientId);
 	}
-	
-	
 }
